added flash support but it freezes

+++ include/usb.h
@@ -12,7 +12,7 @@ void tud_suspend_cb(bool remote_wakeup_en);
 void tud_resume_cb(void);
 
 void tud_hid_report_complete_cb(uint8_t instance, uint8_t const *report,
-                                uint8_t len);
+                                uint16_t len);
 
 uint16_t tud_hid_get_report_cb(uint8_t instance, uint8_t report_id,
                                hid_report_type_t report_type, uint8_t *buffer,
